extract common code: section-document-or-browser
a new file was added before document-or-browser
this was a copy of section-document-or-browser with
many parts stripped out. What remained was mostly common code.

Basic summary:
- switch to using NavTabModelType, so views can be used on the model
- move subTabs into NavTabModel
- extract document-collection-list this renders each section
  of a subTab, in a DocumentCollectionByType element
- extract sub-tabs-panel this renders the Tabs, TabList, and TabPanel.
  the content of the sub tab is customized with a render function
- the css is split between section-document-or-browser
  and sub-tabs-panel
- document-or-browser was renamed to mywork-document-or-browser
  this file had 2 small components. DocumentOrBrowser was not used by
  anything else so, the two components be merged
- document-browser became so simple that it was eliminated
  it's elements were moved into mywork-document-or-browser
