+++ src/components/navigation/section-document-or-browser.tsx
@@ -1,38 +1,26 @@
-import React, { useCallback, useEffect, useMemo } from "react";
+import React, { useEffect } from "react";
 import { observer } from "mobx-react";
 import { useQueryClient } from 'react-query';
-import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import { DocumentModelType } from "../../models/document/document";
 import { getDocumentDisplayTitle } from "../../models/document/document-utils";
 import { logDocumentEvent } from "../../models/document/log-document-event";
-import { NavTabSectionSpec, NavTabSpec } from "../../models/view/nav-tabs";
+import { ISubTabSpec, NavTabModelType } from "../../models/view/nav-tabs";
 import { EditableDocumentContent } from "../document/editable-document-content";
 import { useAppConfig, useClassStore, useProblemStore, useStores,
   useUIStore, useUserStore } from "../../hooks/use-stores";
 import { Logger } from "../../lib/logger";
 import { LogEventName } from "../../lib/logger-types";
 import { useUserContext } from "../../hooks/use-user-context";
-import { DocumentCollectionByType } from "../thumbnail/documents-type-collection";
-import { DocumentDragKey, SupportPublication } from "../../models/document/document-types";
 import { NetworkDocumentsSection } from "./network-documents-section";
 import EditIcon from "../../clue/assets/icons/edit-right-icon.svg";
+import { DocumentCollectionList, kNavItemScale } from "../thumbnail/document-collection-list";
+import { SubTabsPanel } from "./sub-tabs-panel";
 
 import "./section-document-or-browser.sass";
 
-const kNavItemScale = 0.11;
-const kHeaderHeight = 55;
-const kWorkspaceContentMargin = 4;
-const kNavTabHeight = 34;
-const kTabSectionBorderWidth = 2;
-
 interface IProps {
-  tabSpec: NavTabSpec;
+  tabSpec: NavTabModelType;
   isChatOpen?: boolean;
-}
-
-export interface ISubTabSpec {
-  label: string;
-  sections: NavTabSectionSpec[];
 }
 
 export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function SectionDocumentOrBrowser(
@@ -47,32 +35,10 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
   const classStore = useClassStore();
   const navTabSpec = appConfigStore.navTabs.getNavTabSpec(tabSpec.tab);
 
-  const subTabs = useMemo<ISubTabSpec[]>(() => {
-    const _subTabs: ISubTabSpec[] = [];
-    // combine sections with matching titles into a single tab with sub-sections
-    tabSpec.sections?.forEach(section => {
-      const found = _subTabs.findIndex(tab => tab.label === section.title);
-      if (found >= 0) {
-        _subTabs[found].sections.push(section);
-      }
-      else {
-        _subTabs.push({ label: section.title, sections: [section] });
-      }
-    });
-    return _subTabs;
-  }, [tabSpec.sections]);
-
+  const subTabs = tabSpec.subTabs;
   const tabState = navTabSpec && ui.tabs.get(navTabSpec?.tab);
   const _subTabIndex = subTabs.findIndex((subTab) => tabState?.openSubTab === subTab.label);
   const subTabIndex = _subTabIndex < 0 ? 0 : _subTabIndex;
-
-  const hasSubTabs = subTabs.length > 1;
-  const vh = window.innerHeight;
-  const headerOffset = hasSubTabs
-                        ? kHeaderHeight + (2 * (kWorkspaceContentMargin + kNavTabHeight + kTabSectionBorderWidth))
-                        : kHeaderHeight + kNavTabHeight + (2 * (kWorkspaceContentMargin + kTabSectionBorderWidth));
-  const documentsPanelHeight = vh - headerOffset;
-  const documentsPanelStyle = { height: documentsPanelHeight };
 
   useEffect(() => {
     // Set the initial open tab. If the tabSpec changes somehow then the open
@@ -140,55 +106,17 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
     );
   };
 
-  const handleDocumentDragStart = (e: React.DragEvent<HTMLDivElement>, document: DocumentModelType) => {
-    e.dataTransfer.setData(DocumentDragKey, document.key);
-  };
-
-  const handleDocumentStarClick = (document: DocumentModelType) => {
-    document?.toggleUserStar(user.id);
-  };
-
-  const handleDocumentDeleteClick = (document: DocumentModelType) => {
-    ui.confirm("Do you want to delete this?", "Confirm Delete")
-      .then(ok => {
-        if (ok) {
-          document.setProperty("isDeleted", "true");
-          if (document.type === SupportPublication) {
-            logDocumentEvent(LogEventName.DELETE_SUPPORT, { document });
-          }
-        }
-      });
-  };
-
   const renderDocumentBrowserView = (subTab: ISubTabSpec) => {
     const openDocumentKey = tabState?.openDocuments.get(subTab.label);
     const classHash = classStore.classHash;
     return (
       <div>
-        {
-          subTab.sections.map((section: any, index: any) => {
-            const _handleDocumentStarClick = section.showStarsForUser(user)
-              ? handleDocumentStarClick
-              : undefined;
-
-            return (
-              <DocumentCollectionByType
-                key={`${section.type}_${index}`}
-                topTab={navTabSpec?.tab}
-                tab={subTab.label}
-                section={section}
-                index={index}
-                numSections={subTab.sections.length}
-                scale={kNavItemScale}
-                selectedDocument={openDocumentKey}
-                onSelectDocument={handleSelectDocument}
-                onDocumentDragStart={handleDocumentDragStart}
-                onDocumentStarClick={_handleDocumentStarClick}
-                onDocumentDeleteClick={handleDocumentDeleteClick}
-              />
-            );
-          })
-        }
+        <DocumentCollectionList
+          subTab={subTab}
+          tabSpec={tabSpec}
+          selectedDocument={openDocumentKey}
+          onSelectDocument={handleSelectDocument}
+        />
         {
           user.isNetworkedTeacher &&
           <NetworkDocumentsSection
@@ -236,39 +164,14 @@ export const SectionDocumentOrBrowser: React.FC<IProps> = observer(function Sect
   };
 
   return (
-    <div className="document-tab-content">
-      <Tabs
-        className={`document-tabs ${navTabSpec?.tab} ${isChatOpen ? "chat-open" : ""}`}
-        forceRenderTabPanel={true}
-        onSelect={handleTabSelect}
-        selectedIndex={subTabIndex}
-        selectedTabClassName="selected"
-      >
-        <div className={`tab-header-row ${!hasSubTabs ? "no-sub-tabs" : ""}`}>
-          <TabList className={`tab-list ${navTabSpec?.tab}`}>
-            {subTabs.map((subTab) => {
-              const sectionTitle = subTab.label.toLowerCase().replace(' ', '-');
-              const type = subTab.sections[0].type;
-              return (
-                <Tab className={`doc-tab ${navTabSpec?.tab} ${sectionTitle} ${type}`}
-                  key={`section-${sectionTitle}`}>
-                  {subTab.label}
-                </Tab>
-              );
-            })}
-          </TabList>
-        </div>
-        <div className="documents-panel" style={documentsPanelStyle}>
-          {subTabs.map((subTab, index) => {
-            const sectionTitle = subTab.label.toLowerCase().replace(' ', '-');
-            return (
-              <TabPanel key={`subtab-${subTab.label}`} data-test={`subtab-${sectionTitle}`}>
-                { renderDocumentView(subTab) || renderDocumentBrowserView(subTab) }
-              </TabPanel>
-            );
-          })}
-        </div>
-      </Tabs>
-    </div>
+    <SubTabsPanel
+      tabSpec={tabSpec}
+      tabsExtraClassNames={{"chat-open": isChatOpen}}
+      onSelect={handleTabSelect}
+      selectedIndex={subTabIndex}
+      renderSubTabPanel={subTab =>
+        renderDocumentView(subTab) || renderDocumentBrowserView(subTab)
+      }
+    />
   );
 });
