+++ src/components/navigation/sub-tabs-panel.sass
@@ -1,0 +1,88 @@
+@import ../vars
+
+.document-tabs
+  border-color: $charcoal
+  border-width: 0 $tab-section-border-width $tab-section-border-width $tab-section-border-width
+  border-style: solid
+
+  &.my-work
+    border-color: $workspace-teal
+  &.class-work
+    border-color: $classwork-purple
+  &.supports
+    border-color: $support-blue
+  .tab-header-row
+    display: flex
+    flex-direction: row
+    align-items: center
+    background-color: white
+    height: $nav-tab-height
+    padding-top: 3px
+    position: relative
+    &.no-sub-tabs
+      display: none
+  .tab-list
+    margin: 0
+    padding: 0
+    display: flex
+    flex: 1 1 auto
+    &:after
+      content: ""
+      border-bottom: solid 1.5px $charcoal-light-1
+      width: 100%
+      z-index: 1
+      bottom: -2px
+      left: 0
+      position: absolute
+    &.single-panel
+      border-bottom: none
+    .doc-tab
+      flex-grow: 1
+      height: $nav-tab-height
+      display: flex
+      justify-content: center
+      align-items: center
+      cursor: pointer
+      box-sizing: border-box
+      margin: 0 1px
+      border: 1.5px solid $charcoal-light-1
+      border-radius: 5px 5px 0 0
+      background-color: $problem-orange-light-7
+      &:first-of-type
+        margin-left: 2px
+      &:last-of-type
+        margin-right: 0
+      &.my-work
+        background-color: $workspace-teal-light-6
+      &.learning-logs
+        background-color: $learninglog-green-light-7
+      &.class-work
+        background-color: $classwork-purple-light-7
+      &.supports
+        background-color: $support-blue-light-7
+      &:hover
+        background-color: $charcoal-light-9
+        &.my-work
+          background-color: $workspace-teal-light-4
+        &.learning-logs
+          background-color: $learninglog-green-light-5
+        &.class-work
+          background-color: $classwork-purple-light-5
+        &.supports
+          background-color: $support-blue-light-5
+      &:active, &.selected
+        background-color: $charcoal-light-8
+        font-weight: bold
+        border-bottom: none
+        z-index: 2
+        &.my-work
+          background-color: $workspace-teal-light-3
+        &.learning-logs
+          background-color: $learninglog-green-light-4
+        &.class-work
+          background-color: $classwork-purple-light-4
+        &.supports
+          background-color: $support-blue-light-3
+  .documents-panel
+    margin-top: $document-margin
+    overflow-y: auto
